[backend] refacto
